numero,contacto,usuario,rol

+++ src/main/java/upc/com/visiontech2/repositories/RolRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import upc.com.visiontech2.entities.Role;
 
 @Repository
-public interface RolRepository extends JpaRepository<Role, Integer> {
+public interface RolRepository extends JpaRepository<Role, Long> {
 
     
 }
